# cultural-tourism-backend/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + Gin + 腾讯云开发构建的 RESTful API",
        "title": "数字文旅后端 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/pois": {
            "get": {
                "description": "查询点位列表，支持按区域、类型筛选。若传入 lat/lng，结果将包含距离信息(_distance)。",
                "tags": [
                    "POI"
                ],
                "summary": "获取点位列表 (支持LBS)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "区域ID",
                        "name": "region_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "点位类型 (scenic/food/hotel/booth)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "用户纬度 (用于计算距离)",
                        "name": "lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "用户经度 (用于计算距离)",
                        "name": "lng",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的 POI 点位 (景点/饭店/酒店/旅拍机)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POI"
                ],
                "summary": "创建点位",
                "parameters": [
                    {
                        "description": "POI信息",
                        "name": "poi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.POI"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pois/{id}": {
            "get": {
                "tags": [
                    "POI"
                ],
                "summary": "获取点位详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "POI ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.POI"
                        }
                    }
                }
            },
            "put": {
                "description": "更新点位信息 (安全更新，忽略系统字段)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POI"
                ],
                "summary": "更新点位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "POI ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "poi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.POI"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "POI"
                ],
                "summary": "删除点位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "POI ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/themes": {
            "get": {
                "description": "支持按区域筛选。实现PRD“区域优先推荐”：前端应先传region_id查询，若为空则不传region_id查全局。",
                "tags": [
                    "Themes"
                ],
                "summary": "获取主题列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "区域ID",
                        "name": "region_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态 (1:启用)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的旅拍活动主题 (仅管理员)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Themes"
                ],
                "summary": "创建旅拍主题",
                "parameters": [
                    {
                        "description": "主题信息",
                        "name": "theme",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Theme"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/themes/{id}": {
            "get": {
                "tags": [
                    "Themes"
                ],
                "summary": "获取主题详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主题ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "更新主题信息 (支持增量更新)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Themes"
                ],
                "summary": "更新主题",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主题ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "theme",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Theme"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Themes"
                ],
                "summary": "删除主题",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主题ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.POI": {
            "type": "object",
            "properties": {
                "_distance": {
                    "description": "[Audit Fix] 扩展字段：仅用于返回给前端，不存库",
                    "type": "number"
                },
                "_id": {
                    "description": "TCB 自动 ID",
                    "type": "string"
                },
                "_openid": {
                    "description": "[Audit Fix] 系统字段：记录创建者",
                    "type": "string"
                },
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "created_at": {
                    "description": "业务创建时间",
                    "type": "string"
                },
                "desc": {
                    "description": "简介",
                    "type": "string"
                },
                "images": {
                    "description": "轮播图",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "latitude": {
                    "description": "纬度",
                    "type": "number"
                },
                "longitude": {
                    "description": "经度",
                    "type": "number"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "open_time": {
                    "description": "营业时间",
                    "type": "string"
                },
                "phone": {
                    "description": "电话",
                    "type": "string"
                },
                "region_id": {
                    "description": "所属区域",
                    "type": "string"
                },
                "status": {
                    "description": "状态 1:启用 0:禁用",
                    "type": "integer"
                },
                "type": {
                    "description": "类型",
                    "type": "string"
                },
                "updated_at": {
                    "description": "业务更新时间",
                    "type": "string"
                }
            }
        },
...
        "models.Theme": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "TCB 自动生成的 ID",
                    "type": "string"
                },
                "_openid": {
                    "description": "系统字段：创建者",
                    "type": "string"
                },
                "cover": {
                    "description": "封面图 URL",
                    "type": "string"
                },
                "created_at": {
                    "description": "业务创建时间",
                    "type": "string"
                },
                "desc": {
                    "description": "主题简介",
                    "type": "string"
                },
                "name": {
                    "description": "主题名称 (如: \"汉服打卡\")",
                    "type": "string"
                },
                "region_id": {
                    "description": "所属区域 ID (用于实现 PRD 5.1.1 区域优先推荐)",
                    "type": "string"
                },
                "sort": {
                    "description": "排序权重 (默认 100)",
                    "type": "integer"
                },
                "status": {
                    "description": "1:启用 0:禁用",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "业务更新时间",
                    "type": "string"
                }
            }
        }
    }
}